TNO-8 Keylcoak Gold Realm integration (#419)
Refresh Initial DB migration
Add CSS API wrapper library
Add fake CSS API
Add tno-core 0.0.196
Fix user approval process

+++ libs/npm/core/scripts/package.sh
@@ -13,9 +13,15 @@ yarn pack --out ../%s-%v.tgz
 
 # Place the local tgz package in the editor app location.
 echo "Move package ${package} to project"
-mv -f ../${package}.tgz /workspaces/tno/app/editor/
 
-cd /workspaces/tno/app/editor
+# Handle both in devcontainer or out.
+if [ -f "/workspaces/tno/app/editor/" ]; then
+  mv -f ../${package}.tgz /workspaces/tno/app/editor/
+  cd /workspaces/tno/app/editor
+else
+  mv -f ../${package}.tgz ../../../../app/editor/
+  cd ../../../../app/editor
+fi
 
 # Need to do this or it'll ignore additional changes to the same version.
 echo "Clear cache"
